Continue module decomposition: extract auth route handlers
Split auth route concerns into focused response, settings, security, and extension-token modules while keeping route paths and behavior unchanged.

Keep regression safety by preserving existing auth/password/extension flows validated by route tests.
